requirements.txt
streamlit
yt-dlp
pydub
ffmpeg-python
openai-whisper
transformers
torch
gtts
pyannote.audio
torchvision
torchaudio
langdetect
accelerate
sacremoses
reportlab